[==========] 35 tests from 12 test suites ran. (303 ms total)
[  paSSED  ] 34 tests.
[  FAILED  ] 1 test, listed below:
[  FAILED  ] JUNIT.NestedList

+++ src/io/odysz/module/rs.h
@@ -73,7 +73,7 @@ public:
     }
 
     bool next() {
-        if (rowIdx < rows.size())
+        if (rowIdx < static_cast<int>(rows.size()))
             return ++rowIdx < rows.size();
         else return false;
     }
